+++ src/Vector.cpp
@@ -5,25 +5,25 @@
 Vector operator+(const Vector&vec1,const Vector& vec) {
 	return { (vec1.x + vec.x),(vec1.y + vec.y),(vec1.z + vec.z) };
 }
-Vector operator+(const Vector&vec1,const double& vec) {
+Vector operator+(const Vector&vec1,const float& vec) {
 	return { (vec1.x + vec),(vec1.y + vec),(vec1.z + vec) };
 }
 Vector operator*(const Vector&vec1,const Vector& vec) {
 	return { (vec1.x * vec.x),(vec1.y * vec.y),(vec1.z * vec.z) };
 }
-Vector operator*(const Vector&vec1,const double& vec) {
+Vector operator*(const Vector&vec1,const float& vec) {
 	return { (vec1.x * vec),(vec1.y * vec),(vec1.z * vec) };
 }
 Vector operator/(const Vector&vec1,const Vector& vec) {
 	return { (vec1.x / vec.x),(vec1.y / vec.y),(vec1.z / vec.z) };
 }
-Vector operator/(const Vector&vec1,const double& vec) {
+Vector operator/(const Vector&vec1,const float& vec) {
 	return { (vec1.x / vec),(vec1.y / vec),(vec1.z / vec) };
 }
 Vector operator-(const Vector&vec1,const Vector& vec) {
 	return { (vec1.x - vec.x),(vec1.y - vec.y),(vec1.z - vec.z) };
 }
-Vector operator-(const Vector&vec1,const double& vec) {
+Vector operator-(const Vector&vec1,const float& vec) {
 	return { (vec1.x - vec),(vec1.y - vec),(vec1.z - vec) };
 }
 Vector operator-(const Vector& vec1) {
@@ -35,10 +35,10 @@ bool operator==(const Vector& vec1,const Vector& vec) {
 bool operator!=(const Vector& vec, const Vector& vec2) {
 	return ((vec2.x != vec.x) || (vec2.y != vec.y) || (vec2.z != vec.z));
 }
-Vector operator*(const double num, const Vector& vec) {
+Vector operator*(const float num, const Vector& vec) {
 	return { (vec.x * num),(vec.y * num),(vec.z * num) };
 }
-Vector operator/(const double num, const Vector& vec) {
+Vector operator/(const float num, const Vector& vec) {
 	return { (num / vec.x),(num / vec.y),(num / vec.z) };
 }
 inline Vector cross(const Vector& vec1, const Vector& vec2){
@@ -48,10 +48,10 @@ inline Vector cross(const Vector& vec1, const Vector& vec2){
 	cross.z = (vec1.x * vec2.y) - (vec1.y * vec2.x);
 	return cross;
 }
-inline double dot(const Vector& first,const Vector& second) {
+inline float dot(const Vector& first,const Vector& second) {
 	return ((first.x * second.x) + (first.y * second.y) + (first.z * second.z));
 }
-inline double length(const Vector& vec) {
+inline float length(const Vector& vec) {
 	return sqrt((vec.x * vec.x) + (vec.y * vec.y) + (vec.z * vec.z));
 }
 #endif
